pyproject.toml
[project]
name = "4-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.66.3",
    "python-dotenv>=1.0.1",
]
